+++ source/NeuralNetwork.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 NeuralNetwork::~NeuralNetwork(void)
 {
-	for (int i = 0; i<_layers.size(); i++)
+	for (size_t i = 0; i<_layers.size(); i++)
 	{
 		delete _layers[i];
 	}
